Move the `CFFFont` from `src/core/fonts.js` and into its own file
